various improvements (#897)
* [jwt][auth] when refreshing access_token unset from cookies the old refresh_token

* [project] fix casting to float for fps
Fix https://github.com/cgwire/zou/issues/896

+++ zou/app/services/shots_service.py
@@ -1663,7 +1663,7 @@ def set_frames_from_task_type_preview_files(
     project = projects_service.get_project(project_id)
     updates = []
     for shot, preview_duration in results:
-        nb_frames = round(preview_duration * int(project["fps"]))
+        nb_frames = round(preview_duration * float(project["fps"]))
         updates.append(
             {
                 "id": shot.id,
